Cambiado para que no cree siempre el properties

+++ JavaTracer/src/com/javatracer/model/configuration/Configuration.java
@@ -1,4 +1,5 @@
 package com.javatracer.model.configuration;
+import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.FileWriter;
@@ -19,34 +20,39 @@ public class Configuration {
     public Configuration() {
     	fileInput= null;
     	fileWriter =null;
-    	/**
+    	/*
     	 * Create a file configuration
     	 */
 
         this.properties = new Properties();
         try {
-        	/**
+        	/*
         	 * The file is created on disk
         	 */
-        	fileWriter = new FileWriter(CONFIG_FILE_NAME);
-        	/**
-        	 * The file is uploaded to a File to use the Load method
-        	 */
-        	fileInput = new FileInputStream(CONFIG_FILE_NAME);
-        	fileInput = new FileInputStream(CONFIG_FILE_NAME);
-        	/**
-        	 * With the method load, we can load the object properties
-        	 */
-            properties.load(fileInput);
-            /**
-             * Now we add information to the properties through a Hashmap
-             */
-            properties.putAll(getConfig());
-            /**
-             * Save the data on the properties
-             */
-            properties.store(new FileOutputStream("configuration.properties"),"Config");
-            fileWriter.close();
+        	File file = new File(CONFIG_FILE_NAME);
+        	if(!file.exists()){
+        		fileWriter = new FileWriter(CONFIG_FILE_NAME);
+            	
+            	/*
+            	 * The file is uploaded to a File to use the Load method
+            	 */
+            	fileInput = new FileInputStream(CONFIG_FILE_NAME);
+            	
+            	/*
+            	 * With the method load, we can load the object properties
+            	 */
+                properties.load(fileInput);
+                /*
+                 * Now we add information to the properties through a Hashmap
+                 */
+                properties.putAll(getConfig());
+                /*
+                 * Save the data on the properties
+                 */
+                properties.store(new FileOutputStream("configuration.properties"),"Config");
+                fileWriter.close();
+        	}
+        	
         } catch (IOException ex) {
             ex.printStackTrace();
         }
